requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
openai>=1.0.0
pymupdf>=1.24.0
chromadb>=0.5.0
